refactor: streamline external info fetching and improve function naming
